Add ViT QK norm, extended DINOv2 metrics and feature PCA visualization
Added:
  - Add optional QK normalization (qk_norm) support across all ViT variants (ViT/FlexiViT/RoPE ViT/RoPE FlexiViT)
  - Add extended DINOv2 metrics: target entropy + center drift tracking (disable via --no-extended-metrics)
  - Add Feature PCA visualization for feature maps

Fixes:
  - Avoid averaging non-floating state_dict entries in avg-model (e.g. num_batches_tracked, relative_position_index)
  - Fix model config registration to consistently register on the active registry instance

+++ tests_flow/test_train_evaluate_flow.py
@@ -58,7 +58,7 @@ class TestTrainEvaluateFlow(unittest.TestCase):
         self.assertEqual(len(class_to_idx), signature["outputs"][0]["data_shape"][1])
 
         # Average checkpoints
-        avg_models(network, tag, reparameterized=False, epochs=[1, 2], force=True)
+        avg_models(network, tag, reparameterized=False, epochs=[1, 2], accum_dtype=torch.float32, force=True)
 
         # Check average checkpoint is valid
         device = torch.device("cpu")
